Fix timer display granularity issues
Will now force display of the timer to whole seconds and to round up so
that timer does not complete before the strategem is actually ready.
Also updated the ability_hud function to use slimmer math based
calculations that will support expansion to SCH sub and main jobs less
than 99.